Fix unability to save .wav files with samplerate higher that 65535. Previously the samplerate would be truncated, leading to incorrect playback.

+++ waved/wav.d
@@ -230,7 +230,7 @@ void encodeWAV(R)(ref R output, Sound sound) if (isOutputRange!(R, ubyte))
     if (sound.sampleRate == 0)
         throw new WavedException("Can't save a WAV with sample-rate of 0hz. Assign the sampleRate field.");
 
-    output.writeLE!uint(cast(ushort)(sound.sampleRate));
+    output.writeLE!uint(cast(uint)(sound.sampleRate));
 
     size_t bytesPerSec = sound.sampleRate * sound.channels * float.sizeof;
     output.writeLE!uint( cast(uint)(bytesPerSec));
